Rewrite index to be a fully static website (#47)
This update removes backend entirely, and makes site serviceable as a
plain static content. Enables simplified deployment and offline
capabilities through service worker. The data amount is sufficiently
small, this shouldn't cause substantial performance issues

Updated wording to accentuate the main purpose of the project to be
sexpr data files. The site part is a derivative / complementary feature,
not the main point